requirements.txt
termcolor
python-irodsclient
